requirements.txt
python-telegram-bot
-f https://download.pytorch.org/whl/torch_stable.html

fastai==2.1.8
torch==1.7.0+cpu

torchvision==0.8.1+cpu

ipython

 